Initial commit of RegreSQL, a new PostgreSQL tool.
RegreSQL is meant to be a regression testing utility for SQL queries used in
application's code. The idea is that when the queries are maintained as
separate SQL files it's then easy to run them and check their results
against a pre-recorded expected result, or show an error and a diff.

It's inspired from PostgreSQL's regression tests driver, and meant to be
usable for other applications.

Of course pgTap can do the same thing, but you need to somehow upload your
expected result files into your testing database, where we want to maintain
results in our code tree. Also pgTap is more about Unit Tests where the
expected results are a very small result set per query, where here we want
to implement regression testing with potential big results, like a whole
weekly reporting etc.

The idea is to be able to edit queries (for optimisation or readability) and
keep the same behavior as the previous version: regression testing.

This commit only includes the README and some very preliminary code. In go,
because it's high time I play with this language and tooling. Yeah that's
about the excuse.