Review db result handling, implement pretty printing.
Rather than pass around []map[string]interface{} values, implement a proper
ResultSet struct with a separate storage for the column names, allowing to
avoid too much repetition: we won't address the row values by column name in
our use case anyway.

In passing implement a Pretty Printer for our ResultSet, even if we might
end up never using it... sounds a useful to have and it did remind me of my
first programming endeavour (because of the lack of sophistication I guess,
either that of Golang or my usage of it).