Implement a "regresql plan" command.
This allows to re-initialise the query plans when a new query is added,
while skipping already existing plan files. Init does the same and also
creates the setup, including the PostgreSQL connection uri. With "plan" we
keep the existing setup.

In passing centralize the plans, expected and out directories paths in the
Series constructor.